Create  DexTon Entertainments - Codeyoung Final Project
When you access the website, you will notice that you are on the homepage. There is one image containing 4 boxes, which are links to the cartoon page, music page, and so on. If you simply want to use the navbar, there is a dropdown menu and you can click on the page you want to visit! Enjoy!